requirements.txt
albumentations
cycler
matplotlib
numpy
opencv-python
opencv-python-headless
pandas
pillow
safetensors
torch==2.2.2 
torchvision==0.17.2 
torchaudio==2.2.2
torchmetrics
tqdm
transformers
